requirements.txt
streamlit==1.31.1
opencv-contrib-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.3
PySide6==6.6.1
